refactor for main.py
